setup: checking if command_entities table exist
oVirt 3.5 has introduced command_entities table.
In addition to adding some new functions related to this
table to the task clenaer utility, some functions got changed,
but as it is run during setup and might run on top of version which
preceeds oVirt 3.5, it should be checked in the relevant SPs that
the table exists.

Bug-Url: https://bugzilla.redhat.com/1161012

+++ packaging/setup/dbutils/taskcleaner_sp.sql
@@ -31,15 +31,29 @@ CREATE OR REPLACE FUNCTION DeleteAsyncTaskZombiesByCommandId(v_command_id UUID) 
    AS $procedure$
 DECLARE
 deleted_rows int;
+root_command_id_of_deleted_cmds UUID;
 BEGIN
-    IF EXISTS (SELECT 1 FROM GetAsyncTasksZombies() WHERE command_id = v_command_id) THEN
-        DELETE FROM async_tasks WHERE command_id = v_command_id;
+    IF (fn_db_is_table_exists ('command_entities')) THEN
+        DELETE FROM command_entities c WHERE
+        c.command_id IN (
+            SELECT command_id FROM GetAsyncTasksZombies() t WHERE t.command_id = v_command_id
+        );
+
+        DELETE FROM command_entities c WHERE
+        c.command_id IN (
+            SELECT root_command_id FROM GetAsyncTasksZombies() t WHERE t.root_command_id = v_command_id
+        );
     END IF;
-    DELETE FROM command_entities where command_id = v_command_id;
-    GET DIAGNOSTICS deleted_rows = ROW_COUNT;
-    IF deleted_rows > 0 THEN
-        DELETE FROM command_entities C WHERE command_id = root_command_id_of_deleted_cmds AND NOT EXISTS (SELECT * from COMMAND_ENTITIES WHERE root_command_id = C.command_id);
-    END IF;
+
+    DELETE FROM async_tasks WHERE
+    root_command_id IN (
+        SELECT root_command_id FROM GetAsyncTasksZombies() t WHERE t.root_command_id = v_command_id
+    );
+
+    DELETE FROM async_tasks WHERE
+    root_command_id IN (
+        SELECT root_command_id FROM GetAsyncTasksZombies() t WHERE t.command_id = v_command_id
+    );
 END; $procedure$
 LANGUAGE plpgsql;
 
@@ -50,11 +64,13 @@ deleted_rows int;
 root_command_id_of_deleted_cmds UUID;
 BEGIN
         DELETE FROM async_tasks WHERE command_id = v_command_id;
-        SELECT root_command_id  into root_command_id_of_deleted_cmds FROM COMMAND_entities WHERE command_id = v_command_id;
-        DELETE FROM command_entities where command_id = v_command_id;
-        GET DIAGNOSTICS deleted_rows = ROW_COUNT;
-        IF deleted_rows > 0 THEN
+        IF (fn_db_is_table_exists ('command_entities')) THEN
+            SELECT root_command_id  into root_command_id_of_deleted_cmds FROM COMMAND_entities WHERE command_id = v_command_id;
+            DELETE FROM command_entities where command_id = v_command_id;
+            GET DIAGNOSTICS deleted_rows = ROW_COUNT;
+            IF deleted_rows > 0 THEN
                 DELETE FROM command_entities C WHERE command_id = root_command_id_of_deleted_cmds AND NOT EXISTS (SELECT * from COMMAND_ENTITIES WHERE root_command_id = C.command_id);
+            END IF;
         END IF;
 END; $procedure$
 LANGUAGE plpgsql;
